backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

python-multipart
debugpy==1.8.5